+++ sage_trac/ticket_box.py
@@ -6,8 +6,9 @@ from genshi.builder import tag
 from genshi.core import TEXT
 from genshi.filters import Transformer
 
+from trac.cache import cached
 from trac.core import implements, TracError
-from trac.config import Option
+from trac.config import Option, ConfigSection
 from trac.web.api import ITemplateStreamFilter
 from trac.web.chrome import add_stylesheet, ITemplateProvider
 
@@ -18,7 +19,8 @@ from . import git_merger
 import pkg_resources
 import pygit2
 
-FILTER_DATE = Transformer('//div[@id="ticket"]/div[@class="date"]')
+FILTER_PROPERTIES = Transformer(
+    '//div[@id="ticket"]/table[@class="properties"]')
 FILTER_ID = Transformer('//div[@id="ticket"]/h2/a[@class="trac-id"]')
 FILTER_BRANCH = Transformer('//td[@headers="h_branch"]')
 FILTER_BRANCH_TEXT = Transformer('//td[@headers="h_branch"]/text()')
@@ -60,6 +62,35 @@ class TicketBox(git_merger.GitMerger):
             'base URL of the Sage GitLab project from which to show '
             'the ticket build status')
 
+    status_badges_config = ConfigSection('sage_trac:status_badges',
+            """
+            Define custom status badges to display on tickets in this section.
+            Each status badge has a <name> which simply acts as an identity
+            key for the badge, and several options in the format::
+
+                <name>.link_url = <url for anchor tag wrapping the badge>
+                <name>.img_url = <url for the badge image>
+                <name>.order = <relative order of the badge>
+                <name>.height = <height attribute for the image>
+                <name>.width = <width attribute for the image>
+                <name>.margin_left = <CSS margin left of the badge>
+                <name>.margin_left = <CSS margin right of the badge>
+
+            The ``link_url`` and ``img_url`` accept format string templates
+            with the following template variables supported:
+
+            * ``{nonce}``: a random nonce which can be appended to the URL;
+              this can be used to prevent the browser from caching an image
+              for example (this is used e.g. for the patchbot status badges)
+
+            * ``{ticket_<field>}``: for each ticket field there is a variable
+              ``ticket_<field>`` providing its value; e.g. ``ticket_id`` gives
+              the ID of the ticket the badge is displayed on.
+
+            The margin options may be useful for grouping related badges
+            together.
+            """)
+
     # Templates on which this filter should be applied
     _templates = set(['ticket_change.html', 'ticket_preview.html',
                       'ticket.html'])
@@ -74,6 +105,23 @@ class TicketBox(git_merger.GitMerger):
                 'in the "Name <email@example.com>" format')
 
         self._release_signature = pygit2.Signature(m.group(1), m.group(2))
+
+    @cached
+    def status_badges(self):
+        config = self.status_badges_config
+        badges = {}
+        int_opts = set(['order', 'height', 'width'])
+        for option, value in config.options():
+            name, opt = option.split('.', 1)
+            badge = badges.setdefault(name, {'name': name})
+            if opt in int_opts:
+                value = config.getint(option)
+            badge[opt] = value
+
+        def sort_key(b):
+            return ('order' not in b, b.get('order'), b['name'])
+
+        return sorted(badges.values(), key=sort_key)
 
     def filter_stream(self, req, method, filename, stream, data):
         """
@@ -93,22 +141,39 @@ class TicketBox(git_merger.GitMerger):
             FILTER_ID.attr('class', 'trac-id-{0}'.format(ticket['status'])),
         ]
 
-        if self.patchbot_url:
-            # Add the patchbot status icons if a patchbot URL was given
-            nonce = hex(random.randint(0, 1 << 60))
-            ticket_url = '{0}/ticket/{1}'.format(
-                    self.patchbot_url.rstrip('/'), ticket.id)
-            base_url = '{0}/base.svg?nonce={1}'.format(ticket_url, nonce)
-            status_url = '{0}/status.svg?nonce={1}'.format(ticket_url, nonce)
-            elem = tag.div(
-                tag.a(
-                    tag.img(src=base_url, border=0, height=32),
-                    tag.img(src=status_url, border=0, height=32),
-                    href=ticket_url),
-                class_='date')
 
-            filters.append(FILTER_DATE.after(elem))
+        format_vars = {
+            'nonce': hex(random.randint(0, 1 << 60)),
+            'ticket_id': ticket.id
+        }
 
+        for k, v in ticket.values.items():
+            format_vars['ticket_' + k] = v
+
+        badge_tags = []
+        for status_badge in self.status_badges:
+            link_url = status_badge['link_url'].format(**format_vars)
+            img_url = status_badge['img_url'].format(**format_vars)
+            anchor_attrs = {'href': link_url}
+            for anchor_opt in ('margin_left', 'margin_right'):
+                if anchor_opt in status_badge:
+                    anchor_attrs.setdefault('style', '')
+                    anchor_attrs['style'] += (
+                        '{attr}: {val};'.format(
+                            attr=anchor_opt.replace('_', '-'),
+                            val=status_badge[anchor_opt]))
+            img_attrs = {
+                'src': img_url,
+                'border': 0
+            }
+            for img_opt in ('width', 'height'):
+                if img_opt in status_badge:
+                    img_attrs[img_opt] = status_badge[img_opt]
+            badge_tags.append(tag.a(tag.img(**img_attrs), **anchor_attrs))
+
+        filters.append(FILTER_PROPERTIES.after(
+            tag.div(tag.h3('Status badges', id='comment:status-badges'),
+                    tag.div(*badge_tags), class_='badges description')))
         filters.extend(self._get_branch_filters(req, ticket))
 
         def apply_filters(filters):